+++ app.py
@@ -6,9 +6,8 @@ from flask_cors import CORS
 from conf.flask_conf import FlaskConf
 from conf.mail import MailSettings
 from conf.tnlcm import TnlcmSettings
-from core.logs.log_handler import tnlcm_log_handler
-from core.mail.mail import init_mail
 from core.database.database import init_db
+from core.mail.mail import init_mail
 from core.routes import callback_namespace, debug_namespace, library_namespace, trial_network_namespace, user_no_two_factor_namespace, user_namespace, verification_token_namespace
 
 app = Flask(__name__)
@@ -39,4 +38,3 @@ if not MailSettings.TWO_FACTOR_AUTH:
     api.add_namespace(user_no_two_factor_namespace, path="/tnlcm/user")
 else:
     api.add_namespace(verification_token_namespace, path="/tnlcm/verification-token")
-tnlcm_log_handler.info(f"Start Trial Network Lifecycle Manager (TNLCM) {TnlcmSettings.VERSION} on http://0.0.0.0:{TnlcmSettings.TNLCM_PORT}")