+++ core/sites/sites_handler.py
@@ -1,7 +1,7 @@
 from conf.sites import SitesSettings
 from core.repository.repository_handler import RepositoryHandler
 from core.utils.file_handler import load_file, load_yaml, save_yaml
-from core.utils.os_handler import exists_path, is_directory, join_path, list_directory
+from core.utils.os_handler import exists_path, join_path
 from core.utils.parser_handler import ansible_decrypt, yaml_to_dict
 from core.exceptions.exceptions_handler import CustomSitesException
 
@@ -18,8 +18,8 @@ class SitesHandler():
         Constructor
         
         :param https_url: URL of the repository, ``str``
-        :param reference_type: type of reference (branch, tag, commit) to switch, ``str``
-        :param reference_value: value of the reference (branch name, tag name, commit ID) to switch, ``str``
+        :param reference_type: branch to switch, ``str``
+        :param reference_value: value of the branch, ``str``
         :param directory_path: directory path into which the Sites is to be cloned, ``str``
         """
         self.sites_https_url = https_url
@@ -29,15 +29,6 @@ class SitesHandler():
         self.sites_local_directory = join_path(directory_path, self.sites_repository_name)
         self.sites_reference_type = reference_type
         self.sites_reference_value = reference_value
-        if reference_type == "branch" and reference_value:
-            self.sites_reference_value = reference_value
-        elif reference_type == "commit" and reference_value:
-            self.sites_reference_value = reference_value
-        elif reference_type == "tag" and reference_value:
-            self.sites_reference_value = f"refs/tags/{reference_value}"
-        else:
-            self.sites_reference_type = "branch"
-            self.sites_reference_value = SitesSettings.SITES_BRANCH
         self.repository_handler = RepositoryHandler(github_https_url=self.sites_https_url, github_repository_name=self.sites_repository_name, github_local_directory=self.sites_local_directory, github_reference_type=self.sites_reference_type, github_reference_value=self.sites_reference_value)
         self.sites_commit_id = None
     
@@ -76,28 +67,23 @@ class SitesHandler():
         """
         return self.repository_handler.git_tags()
 
-    def get_sites(self) -> list[str]:
+    def validate_site(self, sites_branch: str) -> None:
         """
-        Function to get sites available to deploy trial networks
-
-        :return: list with deployment sites available for deploy trial networks
-        """
-        return [site for site in list_directory(path=self.sites_local_directory) if not site.startswith(".") and is_directory(join_path(self.sites_local_directory, site))]
-
-    def validate_site(self, deployment_site: str) -> None:
-        """
-        Check if deployment site is valid for deploy trial network
+        Check if branch is valid for deploy trial network
         
-        :param deployment_site: trial network deployment site, ``str``
+        :param sites_branch: trial network deployment site, ``str``
         :raise CustomSitesException:
         """
-        if deployment_site not in self.get_sites():
-            raise CustomSitesException(f"Site should be one: {self.get_sites()}", 404)
-        core_file = join_path(self.sites_local_directory, deployment_site, "core.yaml")
+        filtered_branches = [branch for branch in self.git_branches() if branch != "main"]
+        if sites_branch not in filtered_branches:
+            raise CustomSitesException(f"Site should be one: {filtered_branches}", 404)
+        core_file = join_path(self.sites_local_directory, sites_branch, "core.yaml")
+        if not exists_path(path=core_file):
+            raise CustomSitesException(f"File {core_file} not found", 404)
         encrypted_data = load_file(file_path=core_file, mode="rb", encoding=None)
         decrypted_data = ansible_decrypt(encrypted_data=encrypted_data, token=SitesSettings.SITES_TOKEN)
         decrypted_yaml = yaml_to_dict(data=decrypted_data)
-        save_yaml(data=decrypted_yaml, file_path=join_path(self.sites_local_directory, deployment_site, "core_decrypt.yaml"))
+        save_yaml(data=decrypted_yaml, file_path=join_path(self.sites_local_directory, sites_branch, "core_decrypt.yaml"))
 
     def get_site_available_components(self, deployment_site: str) -> dict:
         """